+++ transport/http/proxy.go
@@ -1,4 +1,4 @@
-package transport
+package http
 
 import (
 	"bufio"
