+++ transport/transport.go
@@ -46,11 +46,7 @@ type DialOption func(*DialOptions)
 type ListenOption func(*ListenOptions)
 
 var (
-	DefaultTransport Transport = newHTTPTransport()
+	DefaultTransport Transport
 
 	DefaultDialTimeout = time.Second * 5
 )
-
-func NewTransport(opts ...Option) Transport {
-	return newHTTPTransport(opts...)
-}
